Code cleanup and speeding compile time
