+++ include/MySQL_HostGroups_Manager.h
@@ -46,14 +46,6 @@ class MySrvC;
 class MySrvList;
 class MyHGC;
 
-enum MySerStatus {
-	MYSQL_SERVER_STATUS_ONLINE,
-	MYSQL_SERVER_STATUS_SHUNNED,
-	MYSQL_SERVER_STATUS_OFFLINE_SOFT,
-	MYSQL_SERVER_STATUS_OFFLINE_HARD,
-	MYSQL_SERVER_STATUS_SHUNNED_REPLICATION_LAG
-};
-
 
 std::string gtid_executed_to_string(gtid_set_t& gtid_executed);
 void addGtid(const gtid_t& gtid, gtid_set_t& gtid_executed);
@@ -78,205 +70,9 @@ class GTID_Server_Data {
 	bool readall();
 	bool writeout();
 	bool read_next_gtid();
-/*
-	GTID_Server_Data(struct ev_io *_w, char *_address, uint16_t _port, uint16_t _mysql_port) {
-		active = true;
-		w = _w;
-		size = 1024; // 1KB buffer
-		data = (char *)malloc(size);
-		uuid_server[0] = 0;
-		pos = 0;
-		len = 0;
-		address = strdup(_address);
-		port = _port;
-		mysql_port = _mysql_port;
-	}
-	void resize(size_t _s) {
-		char *data_ = (char *)malloc(_s);
-		memcpy(data_, data, (_s > size ? size : _s));
-		size = _s;
-		free(data);
-		data = data_;
-	}
-	~GTID_Server_Data() {
-		free(address);
-		free(data);
-	}
-	bool readall() {
-		bool ret = true;
-		if (size == len) {
-			// buffer is full, expand
-			resize(len*2);
-		}
-		int rc = 0;
-		rc = read(w->fd,data+len,size-len);
-		if (rc > 0) {
-			len += rc;
-		} else {
-			int myerr = errno;
-			fprintf(stderr,"read returned %d bytes, error %d\n", rc, myerr);
-			if (
-				//(rc == 0) ||
-				(rc==-1 && myerr != EINTR && myerr != EAGAIN)
-			) {
-				ret = false;
-			}
-		}
-		return ret;
-	}
-	bool writeout() {
-		bool ret = true;
-		if (len==0) {
-			return ret;
-		}
-		int rc = 0;
-		rc = write(w->fd,data+pos,len-pos);
-		if (rc > 0) {
-			pos += rc;
-			if (pos >= len/2) {
-				memmove(data,data+pos,len-pos);
-				len = len-pos;
-				pos = 0;
-			}
-		}
-		return ret;
-	}
-	bool read_next_gtid() {
-		if (len==0) {
-			return false;
-		}
-		void *nlp = NULL;
-		nlp = memchr(data+pos,'\n',len-pos);
-		if (nlp == NULL) {
-			return false;
-		}
-		int l = (char *)nlp - (data+pos);
-		char rec_msg[80];
-		if (strncmp(data+pos,(char *)"ST=",3)==0) {
-			// we are reading the bootstrap
-			char *bs = (char *)malloc(l+1-3); // length + 1 (null byte) - 3 (header)
-			memcpy(bs,data+pos+3,l+1-3);
-			char *saveptr1=NULL;
-			char *saveptr2=NULL;
-			//char *saveptr3=NULL;
-			char *token = NULL;
-			char *subtoken = NULL;
-			//char *subtoken2 = NULL;
-			char *str1 = NULL;
-			char *str2 = NULL;
-			//char *str3 = NULL;
-			for (str1 = bs; ; str1 = NULL) {
-				token = strtok_r(str1, ",", &saveptr1);
-				if (token == NULL) {
-					break;
-				}
-				int j = 0;
-				for (str2 = token; ; str2 = NULL) {
-					subtoken = strtok_r(str2, ":", &saveptr2);
-					if (subtoken == NULL) {
-						break;
-					}
-					j++;
-					if (j%2 == 1) { // we are reading the uuid
-						char *p = uuid_server;
-						for (unsigned int k=0; k<strlen(subtoken); k++) {
-							if (subtoken[k]!='-') {
-								*p = subtoken[k];
-								p++;
-							}
-						}
-						//fprintf(stdout,"BS from %s\n", uuid_server);
-					} else { // we are reading the trxids
-						uint64_t trx_from;
-						uint64_t trx_to;
-						sscanf(subtoken,"%lu-%lu",&trx_from,&trx_to);
-						//fprintf(stdout,"BS from %s:%lu-%lu\n", uuid_server, trx_from, trx_to);
-						std::string s = uuid_server;
-						gtid_executed[s].emplace_back(trx_from, trx_to);
-					}
-				}
-			}
-			pos += l+1;
-			free(bs);
-			//return true;
-		} else {
-			strncpy(rec_msg,data+pos,l);
-			pos += l+1;
-			rec_msg[l]=0;
-			//int rc = write(1,data+pos,l+1);
-			//fprintf(stdout,"%s\n", rec_msg);
-			if (rec_msg[0]=='I') {
-				//char rec_uuid[80];
-				uint64_t rec_trxid;
-				char *a = NULL;
-				int ul = 0;
-				switch (rec_msg[1]) {
-					case '1':
-						//sscanf(rec_msg+3,"%s\:%lu",uuid_server,&rec_trxid);
-						a = strchr(rec_msg+3,':');
-						ul = a-rec_msg-3;
-						strncpy(uuid_server,rec_msg+3,ul);
-						uuid_server[ul] = 0;
-						rec_trxid=atoll(a+1);
-						break;
-					case '2':
-						//sscanf(rec_msg+3,"%lu",&rec_trxid);
-						rec_trxid=atoll(rec_msg+3);
-						break;
-					default:
-						break;
-				}
-				fprintf(stdout,"%s:%lu\n", uuid_server, rec_trxid);
-				std::string s = uuid_server;
-				gtid_t new_gtid = std::make_pair(s,rec_trxid);
-				addGtid(new_gtid,gtid_executed);
-				//return true;
-			}
-		}
-		std::cout << "current pos " << gtid_executed_to_string(gtid_executed) << std::endl << std::endl;
-		return true;
-	}
-*/
 	bool gtid_exists(char *gtid_uuid, uint64_t gtid_trxid);
 	void read_all_gtids();
 	void dump();
-/*
-	bool gtid_exists(char *gtid_uuid, uint64_t gtid_trxid) {
-		std::string s = gtid_uuid;
-		auto it = gtid_executed.find(s);
-		fprintf(stderr,"Checking if server %s:%d has GTID %s:%lu ... ", address, port, gtid_uuid, gtid_trxid);
-		if (it == gtid_executed.end()) {
-			fprintf(stderr,"NO\n");
-			return false;
-		}
-		for (auto itr = it->second.begin(); itr != it->second.end(); ++itr) {
-			if ((int64_t)gtid_trxid >= itr->first && (int64_t)gtid_trxid <= itr->second) {
-				fprintf(stderr,"YES\n");
-				return true;
-			}
-		}
-		fprintf(stderr,"NO\n");
-		return false;
-	}
-	void read_all_gtids() {
-		while (read_next_gtid()) {
-		}
-	}
-	void dump() {
-		if (len==0) {
-			return;
-		}
-		read_all_gtids();
-		//int rc = write(1,data+pos,len-pos);
-		fflush(stdout);
-		///pos += rc;
-		if (pos >= len/2) {
-			memmove(data,data+pos,len-pos);
-			len = len-pos;
-			pos = 0;
-		}
-	}
-*/
 };
