+++ include/cpp.h
@@ -2,23 +2,23 @@
 #include "MySQL_Thread.h"
 #include "MySQL_Session.h"
 #include "mysql_backend.h"
-#include "MySQL_Data_Stream.h"
+//#include "MySQL_Data_Stream.h"
 #include "query_cache.hpp"
 #include "mysql_connection.h"
 #include "sqlite3db.h"
 //#include "StatCounters.h"
 #include "MySQL_Monitor.hpp"
-#include "MySQL_Protocol.h"
-#include "MySQL_Authentication.hpp"
-#include "MySQL_LDAP_Authentication.hpp"
+//#include "MySQL_Protocol.h"
+//#include "MySQL_Authentication.hpp"
+//#include "MySQL_LDAP_Authentication.hpp"
 #ifdef PROXYSQLCLICKHOUSE
 #include "ClickHouse_Authentication.hpp"
 #endif /* PROXYSQLCLICKHOUSE */
 #include "fileutils.hpp"
 #include "configfile.hpp"
-#include "query_processor.h"
+//#include "query_processor.h"
 #include "proxysql_admin.h"
-#include "SQLite3_Server.h"
+//#include "SQLite3_Server.h"
 #ifdef PROXYSQLCLICKHOUSE
 #include "ClickHouse_Server.h"
 #endif /* PROXYSQLCLICKHOUSE */
