+++ lib/ClickHouse_Server.cpp
@@ -7,6 +7,9 @@
 #include "proxysql.h"
 #include "cpp.h"
 
+#include "MySQL_Logger.hpp"
+#include "MySQL_Data_Stream.h"
+
 #include <search.h>
 #include <stdlib.h>
 #include <stdio.h>
