+++ lib/MySQL_Logger.cpp
@@ -1,6 +1,9 @@
 #include <fstream>
 #include "proxysql.h"
 #include "cpp.h"
+
+#include "MySQL_Data_Stream.h"
+#include "query_processor.h"
 #include "MySQL_PreparedStatement.h"
 #include "MySQL_Logger.hpp"
