+++ lib/SQLite3_Server.cpp
@@ -7,6 +7,9 @@
 #include "cpp.h"
 
 #include "MySQL_Logger.hpp"
+#include "MySQL_Data_Stream.h"
+#include "query_processor.h"
+#include "SQLite3_Server.h"
 
 #include <search.h>
 #include <stdlib.h>
