+++ lib/mysql_connection.cpp
@@ -4,6 +4,8 @@
 #include <fcntl.h>
 
 #include "MySQL_PreparedStatement.h"
+#include "MySQL_Data_Stream.h"
+#include "query_processor.h"
 
 extern const MARIADB_CHARSET_INFO * proxysql_find_charset_nr(unsigned int nr);
