+++ src/main.cpp
@@ -13,6 +13,11 @@
 #include "MySQL_PreparedStatement.h"
 #include "ProxySQL_Cluster.hpp"
 #include "MySQL_Logger.hpp"
+#include "SQLite3_Server.h"
+#include "query_processor.h"
+#include "MySQL_Authentication.hpp"
+#include "MySQL_LDAP_Authentication.hpp"
+
 
 #include <libdaemon/dfork.h>
 #include <libdaemon/dsignal.h>
